# docs/signaling-openapi.yaml
openapi: 3.0.2
servers:
  - url: https://example.com/signaling/v1
info:
  description: |-
    This specification contains all mandatory and optional endpoints for the [Dataplane Signaling Protocol (DPS)](https://github.com/Metaform/dataplane-signaling/blob/main/docs/signaling.md).
    It is based on OpenAPI 3.0.2 and uses JSON Schema for data modeling.
  version: 0.0.1
  title: Dataplane Signaling Protocol DPS
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
  /dataflows/prepare:
    post:
      tags:
        - Data plane endpoints
      summary: Prepare a DataFlow
      description: |
        The `prepare` request signals to the [=Data Plane=] to initialize a [=Data Flow=] and any resources required for
        data transfer. The request results in a state machine transition to PREPARING or PREPARED. If the state machine
        transitions to PREPARING, the [=Data Plane=] MUST return HTTP 202 Accepted with the `Location` header set to the
        [data flow status relative URL](#status) and a message body containing a `DataFlowResponseMessage`. If the state
        machine transitions to PREPARED, the [=Data Plane=] MUST return HTTP 200 OK and a `DataFlowResponseMessage`.
      operationId: prepareDataFlow
      responses:
        '200':
          description: DataFlow prepared successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataFlowResponseMessage'
              example:
                dataplaneId: dataplane-64345
                state: PREPARED
        '202':
          description: Request received, DataFlow preparation in progress
          headers:
            Location:
              description: URL to poll for DataFlow status
              schema:
                type: string
                format: uri
              example: /dataflows/flow-1234/status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataFlowResponseMessage'
              example:
                dataplaneId: dataplane-64345
                state: PREPARING
        '400':
          description: Bad Request - invalid input, invalid object state or missing required parameters, supplying the
            dataAddress on consumer-pull transfers
      requestBody:
        description: DataFlow preparation request
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DataFlowPrepareMessage'
  /dataflows/{id}/started:
    post:
      tags:
        - Data plane endpoints
      summary: Notify a consumer data plane that a transfer has been started
      description: |
        The `started` request signals to the consumer [=Data Plane=] that a data transmission has begun. For pull
        transfers this indicates the consumer [=Data Plane=] may transmit data. For push transfers this
        indicates the provider has already begun sending data. The request results in a state machine transition to
        STARTED and the Data Plane MUST return HTTP 200 OK and a `DataFlowResponseMessage`.

      operationId: notifyStarted
      parameters:
        - name: id
          in: path
          required: true
          description: The unique identifier of the data flow to start
          schema:
            type: string
            example: flow-1234
      responses:
        '200':
          description: Signals to a data plane that the data transmission has been started normally.
        '400':
          description: Bad Request - invalid input, invalid object state or missing required parameters, omitting the
            dataAddress on consumer-pull transfers or supplying a dataAddress on provider-push transfers.
        '404':
          description: Not Found - data flow with specified ID does not exist
      requestBody:
        description: Notify consumer data plane about started transfer. Required for consumer pull transfers.
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DataFlowStartedNotificationMessage'
  /dataflows/start:
    post:
      tags:
        - Data plane endpoints
      summary: Start a DataFlow (provider side)
      description: |
        The `start` request signals to the [=Data Plane=] to begin a data transfer. The request results in a state
        machine transition to STARTING or STARTED. If the state machine transitions to STARTING, the [=Data Plane=] MUST
        return HTTP 202 Accepted with the `Location` header set to the [data flow status relative URL](#status) and a
        message body containing a `DataFlowResponseMessage`. If the state machine transitions to STARTED, the Data
        Plane MUST return HTTP 200 OK and a `DataFlowResponseMessage`. Repeated DSP transfer starts for existing flows
        (for example, after suspend) MUST be handled by `/dataflows/{id}/resume`.
      operationId: startDataFlow
      requestBody:
        description: DataFlow start request
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DataFlowStartMessage'

      responses:
        '200':
          description: DataFlow started successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataFlowResponseMessage'
              example:
                dataplaneId: dataplane-64345
                state: STARTED
                dataAddress:
                  '@type': DataAddress
                  endpointType: https://23id.org.idsa/v4.1/HTTP
                  endpoint: https://example.com/api/data
                  endpointProperties:
                    - '@type': EndpointProperty
                      name: authorization
                      value: 5up3r53cur3t0k3n
                    - '@type': EndpointProperty
                      name: authType
                      value: bearer
        '202':
          description: Start request received, DataFlow is being started
          headers:
            Location:
              description: The URL to retrieve the data flow status
              schema:
                type: string
                format: uri
              example: /dataflows/flow-1234/status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataFlowResponseMessage'
              example:
                example:
                  dataplaneId: dataplane-64345
                  state: STARTING
        '400':
          description: Bad Request - invalid input, invalid object state or missing required parameters, supplying a
            dataAddress in consumer-pull transfers or omitting the dataAddress in provider-push transfers
        '404':
          description: Not Found - data flow with specified ID does not exist
  /dataflows/{id}/suspend:
    post:
      tags:
        - Data plane endpoints
      summary: Suspend a DataFlow
      description: |
        The suspend request signals to the data plane to suspend an ongoing data transfer.
      operationId: suspendDataFlowById
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DataFlowSuspendMessage'
      parameters:
        - name: id
          in: path
          required: true
          description: The unique identifier of the data flow to suspend
          schema:
            type: string
            example: flow-1234
      responses:
        '200':
          description: DataFlow suspended successfully
        '400':
          description: Bad Request - invalid input, invalid object state or missing required parameters
        '404':
          description: Not Found - data flow with specified ID does not exist
  /dataflows/{id}/resume:
    post:
      tags:
        - Data plane endpoints
      summary: Resume a suspended DataFlow
      description: |
        The resume request signals to the data plane to synchronously resume a suspended data transfer.
        Both the request and response may carry a dataAddress so updated endpoint information can be propagated
        during suspend/start cycles.
      operationId: resumeDataFlowById
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DataFlowResumeMessage'
      parameters:
        - name: id
          in: path
          required: true
          description: The unique identifier of the data flow to resume
          schema:
            type: string
            example: flow-1234
      responses:
        '200':
          description: DataFlow resumed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferStartResponseMessage'
              example:
                dataAddress:
                  '@type': DataAddress
                  endpointType: https://23id.org.idsa/v4.1/HTTP
                  endpoint: https://example.com/api/data
                  endpointProperties:
                    - '@type': EndpointProperty
                      name: authorization
                      value: 5up3r53cur3t0k3n
                    - '@type': EndpointProperty
                      name: authType
                      value: bearer
        '400':
          description: Bad Request - invalid input, invalid object state or missing required parameters
        '404':
          description: Not Found - data flow with specified ID does not exist
  /dataflows/{id}/terminate:
    post:
      tags:
        - Data plane endpoints
      summary: Terminate a DataFlow
      description: |
        The terminate request signals to the data plane to terminate an ongoing data transfer.
      operationId: terminateDataFlowById
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DataFlowTerminateMessage'
      parameters:
        - name: id
          in: path
          required: true
          description: The unique identifier of the data flow to terminate
          schema:
            type: string
            example: flow-1234
      responses:
        '200':
          description: DataFlow terminated successfully
        '400':
          description: Bad Request - invalid input, invalid object state or missing required parameters
        '404':
          description: Not Found - data flow with specified ID does not exist
  /dataflows/{id}/status:
    get:
      tags:
        - Data plane endpoints
      summary: Get the status of a DataFlow
      description: |
        The status endpoint returns a representation of the data flow as managed by the data plane
      operationId: getDataFlowStatusById
      parameters:
        - name: id
          in: path
          required: true
          description: The unique identifier of the data flow to retrieve the status for
          schema:
            type: string
            example: flow-1234
      responses:
        '200':
          description: DataFlow status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataFlowStatusResponseMessage'
              example:
                dataplaneId: dataplane-64345
                state: PREPARING
        '400':
          description: Bad Request - invalid input, invalid object state or missing required parameters
        '404':
          description: Not Found - data flow with specified ID does not exist
  /dataflows/{id}/completed:
    post:
      tags:
        - Data plane endpoints
      summary: Notify the data plane of a completed transfer
      description: |
        The complete endpoint signals to a data plane that the data transmission has completed normally.
      operationId: notifyComplete
      parameters:
        - name: id
          in: path
          required: true
          description: The unique identifier of the data flow to mark as complete
          schema:
            type: string
            example: flow-1234
      responses:
        '200':
          description: DataFlow successfully marked as completed
        '400':
          description: Bad Request - invalid input, invalid object state or missing required parameters
        '404':
          description: Not Found - data flow with specified ID does not exist
...
components:
  schemas:
    DataFlowBaseMessage:
      type: object
      required: [ messageId, participantId, counterPartyId, dataspaceContext, processId, agreementId, datasetId, callbackAddress, transferType, labels, metadata ]
      properties:
        messageId:
          type: string
          description: Unique identifier for the message
          example: msg-123
        participantId:
          type: string
          description: The participant ID of the sender as specified in the Dataspace Protocol
          example: did:web:some-company.com:participants:9cc38260-2644-4f76-8da3-3df675d506a2
        counterPartyId:
          type: string
          description: The participant ID of the counterparty as specified in the Dataspace Protocol
          example: did:web:other-company.com:participants:fc639528-1ece-4164-a49c-a84b59ae18b2
        dataspaceContext:
          type: string
          description: An identifier for the dataspace context when the data plane is used in multiple data spaces.
          example: some-dataspace-context-123
        processId:
          type: string
          description: The transfer process ID as assigned by the control plane for correlation
          example: 545fc6b6-7b1b-47ca-86f7-9ac43c480de6
        agreementId:
          type: string
          description: The contract agreement ID that was negotiated by the control plane
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        datasetId:
          type: string
          description: The ID of the dataset in the DCAT Catalog which is to be transferred
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        callbackAddress:
          type: string
          description: The callback address URL of the control plane where it expects status updates.
          example: https://example.com/controlplanes/xyz/callback
          format: uri
        transferType:
          $ref: '#/components/schemas/TransferType'
        dataAddress:
          $ref: '#/components/schemas/DataAddress'
    DataFlowPrepareMessage:
      allOf:
        - $ref: '#/components/schemas/DataFlowBaseMessage'
    DataFlowStartMessage:
      allOf:
        - $ref: '#/components/schemas/DataFlowBaseMessage'
        - type: object
          required: [ dataFlowId ]
      properties:
        dataAddress:
          $ref: '#/components/schemas/DataAddress'
    DataFlowStartedNotificationMessage:
      type: object
      properties:
        dataAddress:
          $ref: '#/components/schemas/DataAddress'
    DataFlowResumeMessage:
      type: object
      required: [ messageId, processId ]
      properties:
        messageId:
          type: string
          description: Unique identifier for the resume message
          example: b1d5f9e2-3c4b-4f7a-9c3e-2f1e5d6c7b8a
        processId:
          type: string
          description: The transfer process ID as assigned by the control plane for correlation
          example: test-transfer-process-id
        dataAddress:
          $ref: '#/components/schemas/DataAddress'
    TransferStartResponseMessage:
      type: object
      description: |
        Response body used to propagate endpoint updates during transfer start/resume handling.
        The containing object allows adding fields in future versions.
      properties:
        dataAddress:
          $ref: '#/components/schemas/DataAddress'
    DataFlowSuspendMessage:
      type: object
      properties:
        reason:
          type: string
          description: An arbitrary string indicating the reason for suspending the data flow.
          example: "Suspending due to maintenance window."
    DataFlowTerminateMessage:
      type: object
      properties:
        reason:
          type: string
          description: An arbitrary string indicating the reason for terminating the data flow.
          example: "Data source permanently offline."
    DataFlowResponseMessage:
      type: object
      required: [ dataplaneId, state ]
      properties:
        dataplaneId:
          type: string
          description: The unique identifier of the data plane
          example: dataplane-64345
        dataFlowId:
          type: string
          description: The unique identifier of the data flow
          example: dataFlow-71635
        state:
          type: string
          description: The state of the data flow. Only PREPARING and PREPARED are valid in the response to a prepare request.
          enum: [ INITIALIZED, PREPARING, PREPARED, STARTING, STARTED, COMPLETED, SUSPENDED, TERMINATED ]
          example: PREPARED
        dataAddress:
          $ref: '#/components/schemas/DataAddress'
        error:
          type: string
          description: An error message, if any error occurred during the preparation of the data flow.
    DataFlowStatusResponseMessage:
      type: object
      required: [ dataFlowId, state ]
      properties:
        dataFlowId:
          type: string
          description: The unique identifier of the data flow
          example: DataFlow-64345
        state:
          type: string
          description: The current state of the data flow.
          enum: [ INITIALIZED, PREPARING, PREPARED, STARTING, STARTED, COMPLETED, SUSPENDED, TERMINATED ]
    DataAddress:
      type: object
      required:
        - type
        - endpointType
        - endpoint
        - endpointProperties
      properties:
        '@type':
          type: string
          description: The type of data address
          example: DataAddress
        endpointType:
          type: string
          description: the type of endpoint, indicating the protocol to be used
          example: https://23id.org.idsa/v4.1/HTTP
        endpoint:
          type: string
          description: The network location of the endpoint
          example: https://example.com/api/data
          format: uri
        endpointProperties:
          type: array
          description: Additional properties for the endpoint
          items:
            $ref: '#/components/schemas/EndpointProperty'
          example:
            '@type': DataAddress
            endpointType: https://w3.org/idsa/v4.1/HTTP
            endpoint: https://example.com/api/data
            endpointProperties:
              - '@type': EndpointProperty
                name: authorization
                value: 5up3r53cur3t0k3n
              - '@type': EndpointProperty
                name: authType
                value: bearer
    EndpointProperty:
      type: object
      required:
        - '@type'
        - name
        - value
      properties:
        '@type':
          type: string
          default: EndpointProperty
        name:
          type: string
          description: The name of the property
          example: authorization
        value:
          type: string
          description: The value of the property
          example: 5up3r53cur3t0k3n
...
    TransferType:
      type: string
      description: The flow transfer type
      example: com.test.http-pull
...
  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: 'https://petstore.swagger.io/oauth/authorize'
          scopes:
            'write:pets': modify pets in your account
            'read:pets': read your pets
    api_key:
      type: apiKey
      name: api_key
      in: header